+++ src/targets/dom/animate.ts
@@ -1,17 +1,16 @@
 import {
   AcceptedElements,
   AnimationControls,
+  AnimationFactory,
   AnimationListOptions,
-  AnimationWithCommitStyles,
   MotionKeyframesDefinition,
 } from "./types"
 import { animateStyle } from "./animate-style"
 import { getOptions } from "./utils/options"
 import { resolveElements } from "./utils/resolve-elements"
-import { createAnimationControls } from "./utils/controls"
+import { createAnimations } from "./utils/controls"
 import { resolveOption } from "../../utils/stagger"
 import { defaults } from "./utils/defaults"
-import { stopElementAnimation } from "./utils/stop-animation"
 
 export function animate(
   elements: AcceptedElements,
@@ -22,18 +21,9 @@ export function animate(
   const numElements = elements.length
 
   /**
-   * Stop existing animations of all elements in batch before starting
-   * new ones. This will reduce layout thrashing if new animations
-   * need to read current styles.
-   */
-  for (let i = 0; i < numElements; i++) {
-    for (const key in keyframes) stopElementAnimation(elements[i], key)
-  }
-
-  /**
    * Create and start new animations
    */
-  const animations: AnimationWithCommitStyles[] = []
+  const animationFactories: AnimationFactory[] = []
   for (let i = 0; i < numElements; i++) {
     const element = elements[i]
 
@@ -48,13 +38,21 @@ export function animate(
         valueOptions
       )
 
-      animation && animations.push(animation as any)
+      animationFactories.push(animation)
     }
   }
 
-  return createAnimationControls(
-    animations,
-    // TODO: Remove this in case duration is dynamically generated
+  return createAnimations(
+    animationFactories,
+    /**
+     * TODO:
+     * If easing is set to spring or glide, duration will be dynamically
+     * generated. Ideally we would dynamically generate this from
+     * animation.effect.getComputedTiming().duration but this isn't
+     * supported in iOS13 or our number polyfill. Perhaps it's possible
+     * to Proxy animations returned from animateStyle that has duration
+     * as a getter.
+     */
     options.duration ?? defaults.duration
   )
 }