+++ src/targets/dom/data.ts
@@ -5,9 +5,9 @@ const data = new WeakMap<Element, AnimationData>()
 export function getAnimationData(element: Element): AnimationData {
   if (!data.has(element)) {
     data.set(element, {
-      activeTransforms: [],
-      activeAnimations: {},
-      activeGenerators: {},
+      transforms: [],
+      animations: {},
+      generators: {},
       prevGeneratorState: {},
     })
   }