+++ src/targets/dom/types.ts
@@ -3,9 +3,9 @@ import { AnimationGenerator, AnimationGeneratorState } from "../js/types"
 import { NextTime } from "./timeline/types"
 
 export interface AnimationData {
-  activeTransforms: string[]
-  activeAnimations: { [key: string]: BasicAnimationControls | undefined }
-  activeGenerators: { [key: string]: AnimationGenerator | undefined }
+  transforms: string[]
+  animations: { [key: string]: BasicAnimationControls | undefined }
+  generators: { [key: string]: AnimationGenerator | undefined }
   prevGeneratorState: { [key: string]: AnimationGeneratorState | undefined }
 }
 
@@ -165,3 +165,5 @@ export interface CssPropertyDefinition {
 }
 
 export type CssPropertyDefinitionMap = { [key: string]: CssPropertyDefinition }
+
+export type AnimationFactory = () => BasicAnimationControls | undefined