+++ src/targets/dom/utils/apply.ts
@@ -1,9 +1,11 @@
-export const createCssVariableRenderer = (element: Element, name: string) => {
-  return (latest: string | number) =>
-    (element as HTMLElement).style.setProperty(name, latest as string)
-}
+export type RendererFactory = (
+  element: Element,
+  name: string
+) => (latest: string | number) => void
 
-export const createStyleRenderer = (element: Element, name: string) => {
-  return (latest: string | number) =>
-    ((element as HTMLElement).style[name] = latest)
-}
+export const cssVariableRenderer: RendererFactory = (element, name) => (
+  latest
+) => (element as HTMLElement).style.setProperty(name, latest as string)
+
+export const styleRenderer: RendererFactory = (element, name) => (latest) =>
+  ((element as HTMLElement).style[name] = latest)