+++ src/targets/dom/utils/css-var.ts
@@ -1,4 +1,4 @@
-import { transformPropertyDefinitions } from "./transforms"
+import { transformDefinitions } from "./transforms"
 
 export const isCssVar = (name: string) => name.startsWith("--")
 
@@ -6,12 +6,11 @@ export const registeredProperties = new Set<string>()
 
 export function registerCssVariable(name: string) {
   if (registeredProperties.has(name)) return
-
   registeredProperties.add(name)
 
   try {
-    const { syntax, initialValue } = transformPropertyDefinitions.has(name)
-      ? transformPropertyDefinitions.get(name)!
+    const { syntax, initialValue } = transformDefinitions.has(name)
+      ? transformDefinitions.get(name)!
       : ({} as any)
 
     ;(CSS as any).registerProperty({