+++ src/targets/dom/utils/keyframes.ts
@@ -1,14 +1,12 @@
-import { style } from "../style"
 import { UnresolvedValueKeyframe, ValueKeyframe } from "../types"
 
 export function hydrateKeyframes(
   keyframes: UnresolvedValueKeyframe[],
-  element: Element,
-  name: string
+  readInitialValue: () => string | number
 ): ValueKeyframe[] {
   for (let i = 0; i < keyframes.length; i++) {
     if (keyframes[i] === null) {
-      keyframes[i] = i ? keyframes[i - 1] : style.get(element, name)
+      keyframes[i] = i ? keyframes[i - 1] : readInitialValue()
     }
   }