+++ src/targets/dom/utils/options.ts
@@ -8,4 +8,4 @@ export const getOptions = (
   options: AnimationOptionsWithOverrides | AnimationListOptions,
   key: string
 ): AnimationOptions =>
-  options[key] ? { ...options, ...options[key] } : { ...options }
+  options[key] ? { ...options, ...options[key] } : options