+++ src/targets/js/easing/glide/types.ts
@@ -1,8 +1,7 @@
 export interface GlideOptions {
-  // TODO: Look into these namings
   power?: number
-  timeConstant?: number
-  modifyTarget?: (v: number) => number
+  decay?: number
+  changeTarget?: (v: number) => number
 
   max?: number
   min?: number